feat: DevJar react component export
