+++ lib/core.mjs
@@ -1,46 +1,202 @@
-async function createModule(files, { getModulePath }) {
-  let currentImportMap
-  let shim
+import { useEffect, useCallback, useState, useId, useRef } from 'react'
+import { createModule } from './module.mjs'
+import { transform } from 'sucrase'
+import { init, parse } from 'es-module-lexer'
 
-  async function setupImportMap() {
-    if (shim) return shim
-    window.esmsInitOptions = {
-      shimMode: true,
-      mapOverrides: true,
-    }
-    shim = import(/* webpackIgnore: true */ getModulePath('es-module-shims'))
-    await shim
-  }
+let esModuleLexerInit
 
-  function updateImportMap(imports) {
-    imports['react'] = getModulePath('react')
-    imports['react-dom'] = getModulePath('react-dom')
+const isRelative = s => s.startsWith('./')
 
-    const script = document.createElement('script')
-    script.type = 'importmap-shim'
-    script.innerHTML = JSON.stringify({ imports })
-    document.body.appendChild(script)
-    if (currentImportMap) {
-      currentImportMap.parentNode.removeChild(currentImportMap)
-    }
-    currentImportMap = script
-  }
+function transformCode(_code, getModuleUrl, externals) {
+  const code = transform(_code, {
+    transforms: ['jsx', 'typescript'],
+  }).code
 
-
-  function createInlinedModule(code) {
-    return `data:text/javascript;utf-8,${encodeURIComponent(code)}`
-  }
-
-  await setupImportMap()
-  const imports = Object.fromEntries(
-    Object.entries(files).map(([key, code]) => [
-      key,
-      createInlinedModule(code),
-    ])
-  )
-
-  updateImportMap(imports)
-  return self.importShim('index.js')
+  return replaceImports(code, getModuleUrl, externals)
 }
 
-export { createModule }
+function replaceImports(_code, getModuleUrl, externals) {
+  let code = ''
+  let lastIndex = 0
+  let hasReactImports = false
+  const [imports] = parse(_code)
+  imports.forEach(({ s, e, ss, se, n }) => {
+    code += _code.slice(lastIndex, ss)
+    code += _code.substring(ss, s)
+    code += isRelative(n)
+        ? ('@' + n.slice(2))
+        : externals.has(n) ? n : getModuleUrl(n)
+    code += _code.substring(e, se)
+    lastIndex = se
+
+
+    if (n === 'react') {
+      const statement = _code.slice(ss, se)
+      if (statement.includes('React')) {
+        hasReactImports = true
+      }
+    }
+  })
+  code += _code.substring(lastIndex)
+
+  if (!hasReactImports) {
+    code = `import React from 'react';\n${code}`
+  }
+  return code
+}
+
+function createRenderer(createModule_, getModuleUrl) {
+  let reactRoot
+
+  async function render(files) {
+    const mod = await createModule_(files, { getModuleUrl })
+    const React_ = await self.importShim('react')
+    const ReactDOM_ = await self.importShim('react-dom')
+
+    const _jsx = React_.createElement
+    const root = document.getElementById('root')
+    class ErrorBoundary extends React_.Component {
+      constructor(props) {
+        super(props)
+        this.state = { error: null }
+      }
+      componentDidCatch(error) {
+        this.setState({ error })
+      }
+      render() {
+        if (this.state.error) {
+          return _jsx('div', null, this.state.error.message)
+        }
+        return this.props.children
+      }
+    }
+
+    const isReact18 = !!ReactDOM_.createRoot
+    if (isReact18 && !reactRoot) {
+      reactRoot = ReactDOM_.createRoot(root)
+    }
+    const Component = mod.default
+    const element = _jsx(ErrorBoundary, null, _jsx(Component))
+    if (isReact18) {
+      reactRoot.render(element)
+    } else {
+      ReactDOM_.render(element, root)
+    }
+  }
+
+  return render
+}
+
+function createMainScript({ uid }) {
+  const code = (
+`'use strict';
+const _createModule = ${createModule.toString()};
+const _createRenderer = ${createRenderer.toString()};
+
+const getModuleUrl = (m) => window.parent.__devjar__[globalThis.uid].getModuleUrl(m)
+
+globalThis.uid = ${JSON.stringify(uid)};
+globalThis.__render__ = _createRenderer(_createModule, getModuleUrl);
+`)
+  return code
+}
+
+function useLiveCode({ getModuleUrl }) {
+  const iframeRef = useRef()
+  const [error, setError] = useState()
+  const rerender = useState({})[1]
+  const scriptRef = useRef(typeof window !== 'undefined' ? document.createElement('script') : null)
+  const uid = useId()
+
+  // Let getModuleUrl executed on parent window side since it might involve
+  // variables that iframe cannot access.
+  useEffect(() => {
+    if (!globalThis.__devjar__) {
+      globalThis.__devjar__ = {};
+    }
+    globalThis.__devjar__[uid] = {
+      getModuleUrl,
+    }
+
+    return () => {
+      if (globalThis.__devjar__) {
+        delete globalThis.__devjar__[uid]
+      }
+    }
+  }, [])
+
+  useEffect(() => {
+    const iframe = iframeRef.current
+    const doc = iframe && iframe.contentDocument
+
+    if (iframe) {
+      const doc = iframe.contentDocument
+      const div = document.createElement('div')
+      const script = scriptRef.current
+      const scriptContent = createMainScript({ uid })
+
+      div.id = 'root'
+      script.type = 'module'
+      script.id = 'main'
+      script.src = `data:text/javascript;utf-8,${encodeURIComponent(scriptContent)}`
+
+      doc.body.appendChild(div)
+      doc.body.appendChild(script)
+    }
+    return () => {
+      if (iframe) {
+        doc.body.removeChild(doc.getElementById('root'))
+        doc.body.removeChild(doc.getElementById('main'))
+      }
+    }
+  }, [])
+
+  const load = useCallback(async (files) => {
+    if (!esModuleLexerInit) {
+      await init
+      esModuleLexerInit = true
+    }
+
+    if (files) {
+      const overrideExternals =
+        new Set(Object.keys(files).filter(name => !isRelative(name) && name !== 'index.js'))
+
+      // Always share react as externals
+      overrideExternals.add('react')
+      overrideExternals.add('react-dom')
+
+      try {
+        const transformedFiles = Object.keys(files).reduce((res, filename) => {
+          const key = isRelative(filename) ? ('@' + filename.slice(2)) : filename
+          res[key] = transformCode(files[filename], getModuleUrl, overrideExternals)
+          return res
+        }, {})
+
+        const iframe = iframeRef.current
+        const script = scriptRef.current
+        if (iframe) {
+          const render = iframe.contentWindow.__render__
+          if (render) {
+            render(transformedFiles)
+          } else {
+            // if render is not loaded yet, wait until it's loaded
+            script.onload = () => {
+              iframe.contentWindow.__render__(transformedFiles)
+            }
+          }
+        }
+        setError()
+      } catch (e) {
+        setError(e)
+      }
+    }
+    rerender({})
+  }, [])
+
+  return { ref: iframeRef, error, load }
+}
+
+export { 
+  createModule,
+  useLiveCode,
+}
